refactor: put logic of secrets in it's own domain
