+++ src/secret/secret.module.ts
@@ -3,5 +3,6 @@ import { SecretService } from './secret.service';
 
 @Module({
   providers: [SecretService],
+  exports: [SecretService],
 })
 export class SecretModule {}
